+++ simple-stack-flow-masterdetail-fragments/src/main/java/com/example/stackmasterdetailfrag/application/MainActivity.java
@@ -32,6 +32,7 @@ import com.example.stackmasterdetailfrag.util.MasterDetailStateClearStrategy;
 import com.example.stackmasterdetailfrag.util.pathview.HandlesBack;
 import com.example.stackmasterdetailfrag.util.pathview.TabletMasterDetailRoot;
 import com.zhuinden.simplestack.BackstackDelegate;
+import com.zhuinden.simplestack.History;
 import com.zhuinden.simplestack.HistoryBuilder;
 import com.zhuinden.simplestack.StateChange;
 import com.zhuinden.simplestack.StateChanger;
@@ -108,7 +109,7 @@ public class MainActivity
         menu.add("Friends").setShowAsActionFlags(SHOW_AS_ACTION_ALWAYS).setOnMenuItemClickListener(new MenuItem.OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(MenuItem menuItem) {
-                BackstackService.get(MainActivity.this).setHistory(HistoryBuilder.newBuilder() //
+                BackstackService.get(MainActivity.this).setHistory(History.newBuilder() //
                         .add(ConversationListPath.create()) //
                         .add(FriendListPath.create()) //
                         .build(), StateChange.FORWARD);