+++ simple-stack-flow-masterdetail/src/main/java/com/example/stackmasterdetail/application/MainActivity.java
@@ -31,6 +31,7 @@ import com.example.stackmasterdetail.util.BackstackService;
 import com.example.stackmasterdetail.util.MasterDetailStateClearStrategy;
 import com.example.stackmasterdetail.util.pathview.HandlesBack;
 import com.example.stackmasterdetail.util.pathview.TabletMasterDetailRoot;
+import com.zhuinden.simplestack.History;
 import com.zhuinden.simplestack.HistoryBuilder;
 import com.zhuinden.simplestack.StateChange;
 import com.zhuinden.simplestack.StateChanger;
@@ -97,7 +98,7 @@ public class MainActivity
         menu.add("Friends").setShowAsActionFlags(SHOW_AS_ACTION_ALWAYS).setOnMenuItemClickListener(new MenuItem.OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(MenuItem menuItem) {
-                BackstackService.get(MainActivity.this).setHistory(HistoryBuilder.newBuilder() //
+                BackstackService.get(MainActivity.this).setHistory(History.newBuilder() //
                         .add(ConversationListPath.create()) //
                         .add(FriendListPath.create()) //
                         .build(), StateChange.FORWARD);